Moving code to correct location, creating new utility module to help with conversion of dtos.
