+++ libs/shared/modules/src/helpers/mappers/case/communicationStatus.mapper.ts
