+++ libs/shared/modules/src/index.ts
@@ -4,5 +4,5 @@ export * from './case/case.module'
 export * from './health/health.module'
 export * from './journal'
 export * from './journal/models'
-export * from './util/migrations/advert'
+export * from './helpers/migrations/advert'
 export * from './types/result'
