+++ libs/shared/modules/src/utility/utility.service.interface.ts
@@ -1,0 +1,6 @@
+
+
+
+export interface IUtilityService {}
+
+export symbol IUtilityService
